# betterclassroom-backend/swagger.yaml
openapi: 3.1.0
info:
  title: BetterClassroom API
  description: BetterClassroom API documentation
  version: 1.0.0
servers:
  - url: http://betterclassroom.in.htwg-konstanz.de/api
    description: Production server
  - url: http://betterclassroom-staging.in.htwg-konstanz.de:8080/api
    description: Staging server
  - url: http://better-classroom.com:8088/api
    description: Local cluster
  - url: http://localhost:5000/api
    description: Local server
paths:
...
  /students:
    get:
      summary: Returns a list of all students.
      tags:
        - Student
      responses:
        "200":
          description: A JSON array of student objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
        "404":
          description: No students found
    post:
      summary: Add a new student.
      tags:
        - Student
      requestBody:
        description: The student to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
      responses:
        "201":
          description: Student added successfully
        "400":
          description: Student with that ID already exists
        "404":
          description: Given course or exercise not found
    delete:
      summary: Delete all students.
      tags:
        - Student
      responses:
        "200":
          description: All students deleted successfully
  /students/{student_id}:
    get:
      summary: Returns a student by ID.
      tags:
        - Student
      parameters:
        - in: path
          name: student_id
          required: true
          description: ID of the student to return
          schema:
            type: string
      responses:
        "200":
          description: A JSON object of student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student with student ID not found
    delete:
      summary: Delete a student by ID.
      tags:
        - Student
      parameters:
        - in: path
          name: student_id
          required: true
          description: ID of the student to delete
          schema:
            type: string
      responses:
        "200":
          description: Student deleted successfully
        "404":
          description: Student ID not found
  /students/{student_id}/progress:
    get:
      summary: Returns the progress of a student.
      tags:
        - Student
      parameters:
        - in: path
          name: student_id
          required: true
          description: ID of the student whose progress is to be returned
          schema:
            type: string
      responses:
        "200":
          description: The current exercise number of the student
          content:
            application/json:
              schema:
                type: integer
        "404":
          description: Student ID not found
    post:
      summary: Updates the progress of a student.
      tags:
        - Student
      parameters:
        - in: path
          name: student_id
          required: true
          description: ID of the student whose progress is to be updated
          schema:
            type: string
      requestBody:
        description: The progress data to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_exercise:
                  type: integer
      responses:
        "200":
          description: Progress updated successfully
        "400":
          description: No data provided /current_exercise not provided / current_exercise not an integer
        "404":
          description: Student ID not found
  /students/{student_id}/help:
    get:
      summary: Returns the help status of a student.
      tags:
        - Student
      parameters:
        - in: path
          name: student_id
          required: true
          description: ID of the student whose help status is to be returned
          schema:
            type: string
      responses:
        "200":
          description: The help status and table number of the student
          content:
            application/json:
              schema:
                type: object
                properties:
                  help_requested:
                    type: boolean
                  table:
                    type: integer
        "404":
          description: Student ID not found
    post:
      summary: Updates the help status of a student.
      tags:
        - Student
      parameters:
        - in: path
          name: student_id
          required: true
          description: ID of the student whose help status is to be updated
          schema:
            type: string
      responses:
        "200":
          description: Help status updated successfully
        "404":
          description: Student ID not found
...
components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: string
        table:
          type: integer
        course:
          type: string
        exercise:
          type: string
        current_exercise:
          type: integer
        help_requested:
          type: boolean
      required:
        - id
        - table
        - course
        - exercise
      example:
        id: "Ruediger"
        table: 2
        course: "Softwarequalitätssicherung"
        exercise: "Aufgabe 1"
...